+++ app/actions/configure_perspective_correction.py
@@ -5,7 +5,7 @@ import cv2
 from app.components.aruco import ArUco, Marker
 from app.components.camera import Camera
 from app.components.main_loop import MainLoop
-from app.components.ui import UI, UIState
+from app.components.opencv_ui import UI, UIState
 from app.components.helpers import create_dir_if_not_exists
 from app.settings import (ARUCO_DICT, CAMERA, MOCK_IMAGE_PATH,
                           PERSPECTIVE_CORRECTION_MARKER_BUFFER_SIZE, PERSPECTIVE_CORRECTION_MARKER_ID)
