+++ app/components/opencv_ui/text.py
